WIP: fix a few more errors starting up
