+++ lib/plottr_components/src/components/Beamer.js
@@ -8,16 +8,16 @@ import { checkDependencies } from './checkDependencies'
 
 const BeamerConnector = (connector) => {
   const {
-    platform: { openExternal, isWindows, user },
+    platform: { openExternal, isWindows },
   } = connector
 
-  checkDependencies({ openExternal, isWindows, user })
+  checkDependencies({ openExternal, isWindows })
 
-  const paymentId = user.get('payment_id')
-  const customerEmail = user.get('customer_email')
-
-  function Beamer({ inNavigation }) {
+  function Beamer({ inNavigation, user }) {
     const [isInitialized, setInitialized] = useState(false)
+
+    const paymentId = user.payment_id
+    const customerEmail = user.customer_email
 
     const getBeamerAlerts = (num) => {}
 
@@ -64,9 +64,24 @@ const BeamerConnector = (connector) => {
 
   Beamer.propTypes = {
     inNavigation: PropTypes.bool,
+    user: PropTypes.object.isRequired,
   }
 
-  return Beamer
+  const {
+    redux,
+    pltr: { selectors },
+  } = connector
+
+  if (redux) {
+    const { connect } = redux
+    return connect((state) => {
+      return {
+        user: selectors.userSettingsSelector(state.prenent),
+      }
+    })(Beamer)
+  }
+
+  throw new Error('Could not connect Beamer')
 }
 
 export default BeamerConnector
