+++ lib/plottr_components/src/components/containers/DashboardErrorBoundary.js
@@ -17,7 +17,6 @@ const DashboardErrorBoundaryConnector = (connector) => {
       createErrorReport,
       openExternal,
       appVersion,
-      user,
       node: { env },
       rollbar: { rollbarAccessToken, platform },
     },
@@ -27,7 +26,6 @@ const DashboardErrorBoundaryConnector = (connector) => {
     createErrorReport,
     openExternal,
     appVersion,
-    user,
     env,
     rollbarAccessToken,
     platform,
@@ -46,7 +44,14 @@ const DashboardErrorBoundaryConnector = (connector) => {
     }
 
     componentDidMount() {
-      setupRollbar('DashboardErrorBoundary', appVersion, user, env, rollbarAccessToken, platform)
+      setupRollbar(
+        'DashboardErrorBoundary',
+        appVersion,
+        this.props.user,
+        env,
+        rollbarAccessToken,
+        platform
+      )
         .then((rollbar) => {
           this.setState({ rollbar })
         })
@@ -133,6 +138,7 @@ const DashboardErrorBoundaryConnector = (connector) => {
     darkMode: PropTypes.bool,
     proInfo: PropTypes.object,
     trialInfo: PropTypes.object,
+    user: PropTypes.object.isrequired,
   }
 
   const {
@@ -147,6 +153,7 @@ const DashboardErrorBoundaryConnector = (connector) => {
       proInfo: selectors.proInfoSelector(state.present),
       trialInfo: selectors.trialInfoSelector(state.present),
       darkMode: selectors.isDarkModeSelector(state.present),
+      user: selectors.userSettingsSelector(state.present),
     }))(DashboardErrorBoundary)
   }
