+++ lib/plottr_components/src/components/containers/ErrorBoundary.js
@@ -17,7 +17,6 @@ const ErrorBoundaryConnector = (connector) => {
       createErrorReport,
       openExternal,
       appVersion,
-      user,
       node: { env },
       rollbar: { rollbarAccessToken, platform },
     },
@@ -27,7 +26,6 @@ const ErrorBoundaryConnector = (connector) => {
     createErrorReport,
     openExternal,
     appVersion,
-    user,
     env,
     rollbarAccessToken,
     platform,
@@ -46,7 +44,7 @@ const ErrorBoundaryConnector = (connector) => {
     }
 
     componentDidMount() {
-      setupRollbar('ErrorBoundary', appVersion, user, env, rollbarAccessToken, platform)
+      setupRollbar('ErrorBoundary', appVersion, this.props.user, env, rollbarAccessToken, platform)
         .then((rollbar) => {
           this.setState({ rollbar })
         })
@@ -133,6 +131,7 @@ const ErrorBoundaryConnector = (connector) => {
     darkMode: PropTypes.bool,
     proInfo: PropTypes.object,
     trialInfo: PropTypes.object,
+    user: PropTypes.object.isRequired,
   }
 
   const {
@@ -147,9 +146,10 @@ const ErrorBoundaryConnector = (connector) => {
 
     return connect((state) => {
       return {
-        darkMode: state.present.ui.darkMode,
+        darkMode: selectors.isDarkModeSelector(state.present),
         proInfo: selectors.proInfoSelector(state.present),
         trialInfo: selectors.trialInfoSelector(state.present),
+        user: selectors.userSettingsSelector(state.present),
       }
     })(ErrorBoundary)
   }
